% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{gary}
\alias{gary}
\title{Gary is cute, but how cute?}
\usage{
gary(t, h, a, b = 100)
}
\arguments{
\item{t}{is he asking for a treat?}

\item{h}{did he just get a haircut?}

\item{a}{are other animals around?}

\item{b}{baseline cuteness}
}
\value{
cuteness
}
\description{
This is how you find out how cute Gary is
}
