% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm.R
\docType{data}
\name{wm}
\alias{wm}
\title{Average Heights and Weights for American Women}
\format{A data frame with 15 observatins on 2 variables
\itemize{
\item height (inches)
\item weight (lbs)
}}
\source{
The World Almanac and Book of Facts, 1975.
}
\usage{
wm
}
\description{
This data set gives the average heights and weights for American women aged 30–39.
}
\details{
The data set appears to have been taken from the American Society of Actuaries Build and Blood Pressure Study for some (unknown to us) earlier year.
}
\examples{
mean(wm$height)

}
\references{
McNeil, D. R. (1977) Interactive Data Analysis. Wiley.
}
\keyword{datasets}
